Only unsubscribe listeners if destroying activity
Previously, `MainActivity` would unsubscribe all service listeners when
it was stopping (i.e., going to the background). However, that meant
that once it resumed all UI fragments that were sub-classes of
`ServiceAwareFragment` would still reference the old service, which
might have been stopped or replaced, and wouldn't receive any updates if
the service connection is dropped or replaced.