Improve mullvad-exclude
`mullvad-exclude` seems to fail on NixOS with a permissions error when
trying to open /sys/fs/cgroup/net_cls/mullvad-exclude/cgroup.procs. I
noticed that the daemon did not have such issues, so I presumed it's the
options with which the procs file is opened with that are to blame. As
such, I've set the changed the code to use the same options in
`mullvad-exclude` as in `talpid-core`.