Work around banner showing bottom part too soon
If the notification that is to be shown increases the height of the
banner, it will take another layout pass for it to properly adjust the
animation offsets. During this time, the view becomes visible and shows
the bottom part of the banner before the animation starts.

The work around implemented here is to simply delay the animation by a
small amount (enough for a frame to be skipped) to ensure the layout has
completed and moved the banner up enough for the bottom part to not be
shown when the view becomes visible.