Fix return value of `CustomDns.addServer` method
The UI side method returns a boolean flag indicating if the DNS server
was successfully added, which is basically a check if the DNS server
address wasn't already added to the list. Unfortunately, there was a bug
in the method that returned the negated value. This led to a weird bug
where adding a new DNS server caused it to fail first, but succeed on
the second attempt.