Fix `AccountCache` only handled first request
The app would previously crash when the UI side `AccountCache` class
sent a second request message to the service side `AccountCache` class.
The simplest way to reproduce this was to log out and log in again. This
was happening because the service side `AccountCache` actor
implementation was only handling one request and then stopping.

This is now fixed by having the actor run a loop to handle all incoming
requests.