Ensure `ListenerReady` handler is registered first
Previously there could be a race condition because the
`ServiceConnection` constructor would register itself as an event
listener on the service without registering a handler for the
`ListenerReady` event sent after the registration is complete. The
`MainActivity` would then register the handler, and things would work
because the service would send multiple events before sending the
`ListenerReady`, giving time for the handler to be registered before the
event is received.

This commit changes that in order to avoid the race condition. Now the
`ServiceConnection` constructor must receive a callback as a parameter,
and this callback will be used when registering for the `ListenerReady`
event. This allows the constructor to enforce that the handler for the
event is configured before the request for registration is sent.