Add initial logging and telemetry policy
This is nothing new. The app has done exactly this for a long time.
But now it's written down in a policy that's easier for users
and developers to find and understand compared to reading the actual
code. This is added for transparencys sake and for making it easier to
point at during code review.