Fix port confusion when picking WG endpoints
The relay selector was buggy in three ways - when selecting an entry
endpoint for multi-hop, it would consider both TCP and UDP endpoints and
when trying to construct a valid configuration, the endpoint protocol
constraint wouldn't be taken into consideration and finally, the config
construction would assume that only UDP endpoints have been selected if
the transport protocol was not specified. Thus, when using multi-hop the
relay selector could end up picking a UDP configuration with TCP ports.