Add mechanism to reset an account before e2e tests
Adds a mechanism to remove all keys/devices of a specified account in
order to start all tests with a known state. This is achieved using a
simple http client included in the e2e test apk, which support login,
listing devices and removing devices.

Two account used for testing _must_ be specified either as gradle
properties or as runtime arguments to the test runner:
* Local properties (local.properties):
  valid_test_account_token=XXXX
  invalid_test_account_token=XXXX
* Gradle/CLI arguments:
  ./gradlew :e2e:assembleDebug -Pvalid_test_account_token=XXXX -Pinvalid_test_account_token=XXXX
* Runtime arguments:
  am instrument -e valid_test_account_token XXXX -e invalid_test_account_token XXXX # ...