Improve npm alias for running e2e tests against installed app
The tests for an installed version of the app isn't possible to run in
sequence. A lot of them requires the daemon to be in a specific state,
and some require the daemon to change state during the test. This means
that for the tests to be useful they have to be run individually when
the daemon is in the correct state. The `e2e:installed` alias has been
replaced with `e2e:sequential` which can be used to run tests against
the installed app by running `npm run e2e:sequential installed/<test>`.