Update changelog, sync naming in .proto update gui
Updates the changelog, syncs naming to be consistent inside the .proto
file with rust in regards to newly added message types for custom lists.
Update the gui so that it does not break when using a custom list
daemon. However does not let gui actually use custom lists, only has CLI
support.