Change CachedAddresses schema to invalidate cache when updating from 2023.2 to 2023.3
We had a couple of reports pointing at persistent timeouts and SSL errors in API communication using any transport.
The user that happened to be using a broken bridge IP while on 2023.2 would keep using it indefinitely after update to 2023.3 because we used to move the currently used IP to the first position in the list. After update to 2023.3 we would only use that first IP.