Add `mullvad proxy add` command
Add daemon logic for storing custom access methods & allow a user to add
a custom socks5 or shadowsocks proxy.

Add all the necessary information for establishing Socks5
connections (both using a local Socks-proxy as well as the normal,
remote-proxy, use case) and Shadowsocks connections.

Add `api_access_settings` to `mullvad-daemon`

Naturally, the Protobuf types has to be mirrored on the Rust/daemon side
and lots of boilerplate code had to be written to convert between the two.