Do not try to resolve API address if `MULLVAD_API_HOST` is not set
If the `api-override` feature is enabled, we allow overriding the ip
address to the Mullvad API by settings the `MULLVAD_API_ADDRESS` env
variable or have it be resolved automatically if `MULLVAD_API_HOST`
is overriden.

This commit fixes a bug where we would try to resolve the
API address when `MULLVAD_API_ADDRESS` was not set even if
`MULLVAD_API_ADDRESS` was not set either.