Properly pass arguments to Cargo in `build-apk.sh`
The `build-apk.sh` script broke in 8b4d5aa7ba4cdb7a3f45ce271b4e7ee174c75882
due to quoting a variable containing a string with space seperated
arguments. While it is best practice to quote variables to prevent word
splitting, word splitting was the behavior that we wanted in this
particular case.

This commit fixes this issue by building an array of arguments instead,
which when expanded will pass along the arguments to `cargo build` properly
as multiple strings. The script now works, and `shellcheck` is happy!