Fix false negative in github action
Add a precondition to all os-specific test steps in the desktop
end-to-end test github action which checks if there are any vms to test
for the given os. . Without this check, the parsing of the vm names
assumed that there would exist atleast 1, causing the parsing to fail if
there were none. This would show up in the GitHub summary view as a
failure, while the intention was for it to be a noop.

Fix `yamllint`