Fix bridge interface IP on macOS
The IP of the bridge interface was incorrectly
assumed to be constant, but has at some point
changed. This broke the algorithm for finding
the bridge interface.

Replace the algorithm with one based on the
guest IP, which is given by `tart`. As the bridge
IP is no longer constant, we also expose it through
the `TEST_CONFIG`.