Build Linux app in release mode for end-to-end tests
We have observed the "download app" step taking a considerable amount of
time. Since we are testing so many Linux platforms, it is worth spending
the extra time up front compiling the app in release mode to cut the
download time, since we are building the app once but the number of
downloads scales with the number of tested platforms.